added command to create versions.yml
